backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openai==1.12.0
google-generativeai==0.3.2
httpx==0.26.0
sqlalchemy==2.0.27
redis==5.0.1
chromadb==0.4.22
jinja2==3.1.3
python-slugify==8.0.4
transitions==0.9.0
